requirements.txt
easyocr
opencv-python-headless
numpy
pillow
gradio
